Switch ingestion from Tradier to Deribit
Tradier requires US brokerage eligibility, which isn't available here.
Deribit's market data API is fully public (no account/API key), reuniting
the code with what CLAUDE.md/README described from the start.

- DeribitClient replaces TradierClient: get_instruments +
  get_book_summary_by_currency, converts crypto-denominated bid/ask/last
  to USD via each instrument's index price.
- No Settings/auth needed, so pydantic-settings and .env.example are gone.
- OptionQuote.volume/open_interest are now floats (fractional crypto
  contract sizes).
- SPEC.md updated: crypto framing, European-only vol solver (Deribit
  options don't have American exercise), 24/7 schedule instead of
  market-hours.

+++ src/vollab/ingestion/models.py
@@ -25,8 +25,8 @@ class OptionQuote(BaseModel):
     bid: float = Field(ge=0)
     ask: float
     last: float | None = None
-    volume: int
-    open_interest: int
+    volume: float
+    open_interest: float
     underlying_price: float
 
     @field_validator("snapshot_ts")
